+++ KiKickBoard/KiKickBoard/Info.plist
@@ -2,6 +2,12 @@
 <!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
 <plist version="1.0">
 <dict>
+	<key>KAKAO_APP_KEY</key>
+	<string>db03961a6321c7b38df7a5234d5aa393
+db03961a6321c7b38df7a5234d5aa393
+</string>
+	<key>KAKAO_PHASE</key>
+	<string>&quot;sandbox&quot;</string>
 	<key>UIApplicationSceneManifest</key>
 	<dict>
 		<key>UIApplicationSupportsMultipleScenes</key>
